+++ src/main/java/util/FileFormatter.java
@@ -1,3 +1,7 @@
+package util;
+
+import tasks.Task;
+
 import java.util.List;
 
 public class FileFormatter {
