+++ src/main/java/util/FileParser.java
@@ -1,5 +1,11 @@
-import AlanExceptions.AlanException;
-import AlanExceptions.FileCorruptException;
+package util;
+
+import alanExceptions.AlanException;
+import alanExceptions.FileCorruptException;
+import tasks.Deadline;
+import tasks.Event;
+import tasks.Task;
+import tasks.Todo;
 
 import java.util.ArrayList;
 import java.util.List;
