+++ src/main/java/util/ParsedData.java
@@ -1,3 +1,5 @@
+package util;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
